Feat/order

+++ src/main/java/com/example/fashion_commerce/admin/responses/OrderAnalytics.java
@@ -1,8 +1,8 @@
 package com.example.fashion_commerce.admin.responses;
 
 public record OrderAnalytics(
-    int totalOrder,
-    int pendingOrders,
-    int deliveredOrders,
-    int cancelledOrders
+    double totalOrder,
+    double pendingOrders,
+    double deliveredOrders,
+    double cancelledOrders
 ) { }
